requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.0
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
beautifulsoup4=4.13.4=pypi_0
bs4=0.0.2=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.6.15=hbd8a1cb_0
certifi=2025.6.15=pypi_0
charset-normalizer=3.4.2=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
dataset-builder-inat=1.0.4=pypi_0
filelock=3.13.1=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.58.4=pypi_0
fsspec=2024.6.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.5.1=pypi_0
kiwisolver=1.4.8=pypi_0
ld_impl_linux-64=2.43=h1423503_5
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgomp=15.1.0=h767d61c_2
liblzma=5.8.1=hb9d3cd8_2
libmpdec=4.0.0=hb9d3cd8_0
libsqlite=3.50.1=hee588c1_0
libuuid=2.38.1=h0b41bf4_0
libzlib=1.3.1=hb9d3cd8_2
markupsafe=3.0.2=pypi_0
matplotlib=3.10.3=pypi_0
matplotlib-venn=1.1.2=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h2d0b736_3
networkx=3.3=pypi_0
numpy=2.1.2=pypi_0
nvidia-cublas-cu12=12.8.3.14=pypi_0
nvidia-cuda-cupti-cu12=12.8.57=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.61=pypi_0
nvidia-cuda-runtime-cu12=12.8.57=pypi_0
nvidia-cudnn-cu12=9.7.1.26=pypi_0
nvidia-cufft-cu12=11.3.3.41=pypi_0
nvidia-cufile-cu12=1.13.0.11=pypi_0
nvidia-curand-cu12=10.3.9.55=pypi_0
nvidia-cusolver-cu12=11.7.2.55=pypi_0
nvidia-cusparse-cu12=12.5.7.53=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.8.61=pypi_0
nvidia-nvtx-cu12=12.8.55=pypi_0
onnx=1.18.0=pypi_0
onnxruntime=1.22.0=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=3.5.0=h7b32b05_1
packaging=25.0=pypi_0
pandas=2.3.0=pypi_0
pillow=11.0.0=pypi_0
pip=25.1.1=pyh145f28c_0
protobuf=6.31.1=pypi_0
pyarrow=20.0.0=pypi_0
pyparsing=3.2.3=pypi_0
python=3.13.5=hec9711d_102_cp313
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.13=7_cp313
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h8c095d6_2
requests=2.32.4=pypi_0
scikit-learn=1.7.0=pypi_0
scipy=1.15.3=pypi_0
setuptools=70.2.0=pypi_0
six=1.17.0=pypi_0
soupsieve=2.7=pypi_0
sympy=1.13.3=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.13=noxft_hd72426e_102
torch=2.7.1+cu128=pypi_0
torchaudio=2.7.1+cu128=pypi_0
torchvision=0.22.1+cu128=pypi_0
tqdm=4.67.1=pypi_0
triton=3.3.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pypi_0
